@@ -4,11 +4,11 @@ source "https://rubygems.org/"
 
 gem "dotenv", require: "dotenv/load"
 
-gem "rails", "~> 7.2.0"
+gem "rails", "~> 8.0.0"
 gem "pg"
 gem "dalli", platforms: :ruby
 gem "simple_form"
-gem "active_model_serializers", "~> 0.10.0"
+gem "blueprinter"
 gem "ruby-vips"
 gem "bcrypt", require: "bcrypt"
 gem "draper"
@@ -44,6 +44,8 @@ gem "datadog", require: "datadog/auto_instrument"
 gem "opensearch-ruby"
 
 gem "mailgun-ruby"
+gem "mjml-rails"
+gem "mrml"
 
 gem "faraday"
 gem "faraday-follow_redirects"
