@@ -2,14 +2,13 @@
 
 module Maintenance
   module User
-    class EmailConfirmationMailer < ActionMailer::Base
-      helper ApplicationHelper
-      helper UsersHelper
-      default :from => Danbooru.config.mail_from_addr, :content_type => "text/html"
-
+    class EmailConfirmationMailer < ApplicationMailer
       def confirmation(user)
         @user = user
-        mail(:to => @user.email, :subject => "#{Danbooru.config.app_name} account confirmation")
+        mail(
+          to: user_email(@user),
+          subject: "#{Danbooru.config.app_name} Account Confirmation",
+        )
       end
     end
   end
