@@ -2,14 +2,14 @@
 
 module Maintenance
   module User
-    class LoginReminderMailer < ActionMailer::Base
-      default :from => Danbooru.config.mail_from_addr, :content_type => "text/html"
-
+    class LoginReminderMailer < ApplicationMailer
       def notice(user)
         @user = user
-        if user.email.present?
-          mail(:to => user.email, :subject => "#{Danbooru.config.app_name} login reminder")
-        end
+        return if user.email.blank?
+        mail(
+          to: user_email(@user),
+          subject: "#{Danbooru.config.app_name} Login Reminder",
+        )
       end
     end
   end
