@@ -5,11 +5,11 @@ class PostEventsController < ApplicationController
 
   def index
     @events = PostEventDecorator.decorate_collection(
-      PostEvent.includes(:creator).search(search_params).paginate(params[:page], limit: params[:limit])
+      PostEvent.includes(:creator).search(search_params).paginate(params[:page], limit: params[:limit]),
     )
     respond_with(@events) do |format|
       format.json do
-        render json: Draper.undecorate(@events)
+        render json: { post_events: PostEventBlueprint.render_as_hash(Draper.undecorate(@events)) }
       end
     end
   end
