@@ -144,7 +144,11 @@ module ParseValue
       return ago if ago.present?
 
       begin
-        Time.zone.parse(object)
+        parsed_date = Time.zone.parse(object)
+
+        # OpenSearch's strict_date_optional_time format only supports years 0-9999
+        return nil if parsed_date && (parsed_date.year < 0 || parsed_date.year > 9999)
+        parsed_date
       rescue ArgumentError
         nil
       end
