@@ -1,8 +1,6 @@
 # frozen_string_literal: true
 
 class PostPruner
-  DELETION_WINDOW = 30
-
   def prune!
     Post.without_timeout do
       prune_pending!
@@ -12,9 +10,11 @@ class PostPruner
   protected
 
   def prune_pending!
+    window = Danbooru.config.unapproved_post_deletion_window
+
     CurrentUser.as_system do
-      Post.where("is_deleted = ? and is_pending = ? and created_at < ?", false, true, DELETION_WINDOW.days.ago).find_each do |post|
-        post.delete!("Unapproved in #{DELETION_WINDOW} days")
+      Post.where("is_deleted = ? and is_pending = ? and created_at < ?", false, true, window.ago).find_each do |post|
+        post.delete!("Unapproved in #{window.inspect}")
       rescue PostFlag::Error
         # swallow
       end
