@@ -1,13 +1,15 @@
 # frozen_string_literal: true
 
 class RelatedTagQuery
-  include ActiveModel::Serializers::JSON
-
   attr_reader :query, :category_id
 
   def initialize(query: nil, category_id: nil)
     @query = TagAlias.to_aliased(query.to_s.downcase.strip).join(" ")
     @category_id = category_id
+  end
+
+  def as_json(_options = {})
+    serializable_hash
   end
 
   def tags
