@@ -2,13 +2,17 @@
 
 class TagCorrection
   include ActiveModel::Model
-  include ActiveModel::Serializers::JSON
 
   attr_reader :tag
+
   delegate :category, :post_count, :real_post_count, to: :tag
 
   def initialize(tag_id)
     @tag = Tag.find(tag_id)
+  end
+
+  def as_json(_options = {})
+    attributes
   end
 
   def attributes
